refactor: rename utils folder to common
