+++ src/fs/common/check-exist.js
