+++ src/fs/list.js
@@ -1,6 +1,6 @@
 import * as fs from 'node:fs/promises';
-import isExist from './utils/check-exist.js';
-import { SOURCE_FOLDER_RELATIVE_PATH } from './utils/config.js';
+import isExist from './common/check-exist.js';
+import { SOURCE_FOLDER_RELATIVE_PATH } from './common/config.js';
 
 const list = async () => {
     const folderPath = new URL(SOURCE_FOLDER_RELATIVE_PATH, import.meta.url);
